chore(training): Improved error handling for incorrect input
