+++ src/app/services/__init__.py
@@ -1,8 +1,10 @@
+from .exceptions import DataValidationError
 from .helper import load_model, save_model
 from .prediction import NoTrainedModelError, PredictionService
 from .training import TrainingService
 
 __all__ = [
+    "DataValidationError",
     "NoTrainedModelError",
     "PredictionService",
     "TrainingService",
