+++ src/app/services/exceptions.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class DataValidationError(Exception):
+    message: str
