+++ src/app/utils/__init__.py
@@ -1,4 +1,3 @@
-from .df_utils import get_feats_and_target
 from .exampler import ExamplerMixIn
 
-__all__ = ["ExamplerMixIn", "get_feats_and_target"]
+__all__ = ["ExamplerMixIn"]
